Delete key_strike file
-Remove key_strike file to solve unresetable points_label bug
